Contest 2 Task Q
